Update plugin tests to new style
