+++ test/integration/CMakeLists.txt
@@ -3,8 +3,30 @@
 # and RAJA project contributors. See the RAJA/COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
-################################################################################
+###############################################################################
 
-raja_add_test(
-  NAME test-plugin
-  SOURCES test_plugin.cpp plugin_for_test.cpp)
+list(APPEND PLUGIN_BACKENDS Sequential)
+
+if(RAJA_ENABLE_OPENMP)
+  list(APPEND PLUGIN_BACKENDS OpenMP)
+endif()
+
+if(RAJA_ENABLE_TBB)
+  list(APPEND PLUGIN_BACKENDS TBB)
+endif()
+
+if(RAJA_ENABLE_CUDA)
+  list(APPEND PLUGIN_BACKENDS Cuda)
+endif()
+
+if(RAJA_ENABLE_HIP)
+  list(APPEND PLUGIN_BACKENDS Hip)
+endif()
+
+if(RAJA_ENABLE_TARGET_OPENMP)
+  list(APPEND PLUGIN_BACKENDS OpenMPTarget)
+endif()
+
+add_subdirectory(plugin)
+
+unset( PLUGIN_BACKENDS )
