+++ test/integration/plugin/tests/counter.hpp
